Skybox and Camera Experimentation
add: several auto-generated skyboxes
add: several experimental camera features
fix: open job and work menu only for available structures

(Ugly scene)

+++ Assets/_Scripts/Structs/Goals.cs
@@ -3,9 +3,9 @@ namespace BWV
 {
     public struct Goals
     {
-        public int gold;
-        public int happiness;
-        public int favor;
-        public int title;
+        public float gold;
+        public float happiness;
+        public float favor;
+        public float title;
     }
 }
